requirements.txt
sniffer
click
expecter
pytest
pexpect
cx_Oracle
requests
selenium
Appium-Python-Client
paramiko
scp
jira
ConfigParser
pyhamcrest
pymysql
pytest-rerunfailures
pytest-html
sphinx